Header and main content edits
Changed sidebar color to #bc5090
Changed header items to be flex layout with 2 rows.
Changed main content to be 2 rows and 2 columns in order to include the "Your projects", "Announcements" and "Trending" headings
Projects column is grid and second column is flex for announcements and trending cards
Added all the dummy HTML text for the whole page.